+++ src/client/components/010-app.js
@@ -97,6 +97,7 @@ class App extends React.Component {
         onRedirectToRoot={() => { this.setState({ fRedirectToRoot: true }); }}
         fRedirectToRoot={this.state.fRedirectToRoot}
         query={this.props.location.query}
+        saveAsBaseline={this.saveAsBaseline}
       />
     );
   }
@@ -150,6 +151,17 @@ class App extends React.Component {
     .catch(() => {
       // $FlowFixMe
       this.setState({ error: `Could not find ${filePath}` });
+    });
+  }
+
+  saveAsBaseline = (snapshotId: string) => {
+    fetch('/api/saveAsBaseline', {
+      method: 'POST',
+      headers: { 'content-type': 'application/json' },
+      body: JSON.stringify({
+        filePath: this.state.fetchedItemPath,
+        id: snapshotId,
+      }),
     });
   }
 }